Ran vs code cleanup

+++ SwanSong.Helper/PagingHelper.cs
@@ -10,9 +10,9 @@ public class PagingHelper
     public static PagedResponse<List<T>> CreatePagedReponse<T>(List<T> pagedData, PaginationFilter validFilter, long totalRecords)
     {
         var respose = new PagedResponse<List<T>>(pagedData, validFilter.PageNumber, validFilter.PageSize);
-        var totalPages = ((double)totalRecords / (double)validFilter.PageSize);
+        var totalPages = (totalRecords / (double)validFilter.PageSize);
         int roundedTotalPages = Convert.ToInt32(Math.Ceiling(totalPages));
-        
+
         respose.TotalPages = roundedTotalPages;
         respose.TotalRecords = totalRecords;
